Enhance chat functionality and improve flashcards handling in chainlit_frontend
- Switched to GPT-4o model for improved chat interactions.
- Added shutil import for directory management.
- Implemented user feedback upon successful notebook upload.
- Transitioned from print statements to logger for consistent logging.
- Increased recursion limit for message processing to 10.
- Streamlined flashcards creation process, including automatic file discovery and cleanup.
- Added chat end handler to clean up the flashcards directory, ensuring a fresh start for each session.
