+++ .gitignore
@@ -82,10 +82,16 @@ target/
 profile_default/
 ipython_config.py
 
+# FLashcard directory
+flashcards/
+
+# .chainlit directory
+.chainlit/
+
 # pyenv
 #   For a library or package, you might want to ignore these files since the code is
 #   intended to run in multiple environments; otherwise, check them in:
-# .python-version
+.python-version
 
 # pipenv
 #   According to pypa/pipenv#598, it is recommended to include Pipfile.lock in version control.
