+++ flashcards_cca7854c-91c2-47d5-872f-46132739ace0.csv
@@ -1,0 +1,11 @@
+Front,Back
+What command is used to clone a GitHub repository in a notebook?,!git clone https://github.com/arcee-ai/DALM
+How do you install or upgrade a Python package in a notebook?,!pip install --upgrade -q -e .
+Which command installs the 'langchain' and 'langchain-community' libraries?,!pip install -qU langchain langchain-core langchain-community sentence_transformers
+What is the command to install 'pymupdf' and 'faiss-cpu'?,!pip install -qU pymupdf faiss-cpu
+How do you import the Pandas library in Python?,import pandas as pd
+Which library provides the 'HuggingFaceEmbeddings' class?,from langchain_community.embeddings import HuggingFaceEmbeddings
+How do you import the 'FAISS' vector store from the 'langchain_community' library?,from langchain_community.vectorstores import FAISS
+What is the import statement for reading directories using the 'Llama Index' library?,from llama_index.core import SimpleDirectoryReader
+Which import statement is used for parsing nodes in the 'Llama Index' library?,from llama_index.core.node_parser import SimpleNodeParser
+How do you import the 'MetadataMode' schema from the 'Llama Index' library?,from llama_index.core.schema import MetadataMode
